FIX
I have to fix many things like my read to fit the assignment standards and have to implement the nonprintable ASCII values. More changes to come to the logic behind the decimal conversion method. Also, I have to make sure to check that the input can function the way the professor wanted it to.